update: fix req fields

+++ app/Common/Models/Gallery/HasGalleryImage.php
@@ -7,16 +7,19 @@ namespace App\Common\Models\Gallery;
  */
 trait HasGalleryImage
 {
-    public function setImage(string $image): static
+    public function setImage(?string $image): static
     {
-        $post['image'] = $image;
-        $post['images_path'] = $this->setImagesPath();
-        if ($this->image) {
-            unlink(public_path($this->image));
+        if( !empty($image)) {
+            $post['image'] = $image;
+            $post['images_path'] = $this->setImagesPath();
+            if($this->image) {
+                unlink(public_path($this->image));
+            }
+            if($urlImage = GalleryImage::uploadSingleImage($post)) {
+                $this->image = $urlImage;
+            }
         }
-        if ($urlImage = GalleryImage::uploadSingleImage($post)) {
-            $this->image = $urlImage;
-        }
+
         return $this;
     }
